mangrove.js: Improve detection of read-only mode
JsonRpcProvider implements getSigner, but does not check that the signer is valid when the method is called. So in case the signer is invalid, mangrove.js fails with an Error.
The detection logic has therefore been extended to check that getAddress() on the signer does not fail; If it does, don't use that signer.