mangrove.js: Semibook cache now tracks whether the offer list is empty
This fixes an issue where if the cache was empty, `getPivot` would always assume the offer list was empty.
By caching the id of the best offer (regardless of whether the offer is in the cache) we can determine whether there are actually offers outside the cache that must be taken into consideration.