mangrove.js: Make Semibook.getPivotId fetch offer on-demand
`Semibook.getPivotId` will now fetch offers if the cache is insufficient to find a pivot. Fetched offers will be put in the cache, but only up to the specified `maxOffers` parameter.

In order to ensure cache consistency, Semibook now uses a mutex to lock the cache during updates.

It now also updates on "block" events instead of on individual events to avoid interleaving updates of the cache on the same block. This could happen if a block contained multiple events, and an event listener on the Semibook called `getPivotId` and amended the cache before all of the events in the block had been processed.