build: Consistently handle package working directories in the CI workflow
For historical reasons, the default working directory was `packages/mangrove-solidity` and the jobs for that package therefore didn't specify a working package.
However, this adds unnecessary complexity, resulting in issues like the one fixed in @556e3af.
This commit therefore removes the default working directory and makes it explicit for the `mangrove-solidity` jobs.