build: Add a Heroku-specific build that avoids Foundry (#564)
* build: Add a Heroku-specific build path that avoids Forge

Temp workaround for the broken build on Heroku

* build: Don't build mangrove-solidity on Heroku

The ABI's are needed but these are currently being committed to the src folder of mangrove.js.

* build: Don't generate hardhatAddresses.json on Heroku

Generation of hardhatAddresses.json is broken on Heroku due to mangrove-solidity build being disabled. However, tests aren't run on Heroku, so hardhatAddresses.json isn't needed.