fix!: do not poll logevents on every block
Previously, semibooks listened for new blocks,
then queried synchronously for events for that block,
and then updated according to those events.

That processing was inefficient, and became very impractical
over a websocket - where events only arrive one-by-one (but in order).

semibook and market have been rewritten to remove subscriptions and waits
for the processing of events by block. Using code have been rewritten to listen
for logevents, instead.

Test and test-utils have been rewritten to rely on logevents, as well.

Functionality and API in market and semibook supporting on-block listening has been
removed as well as supporting code.