ci: Add missing permission
dorny test-reporter now requires the 'checks' write permission, see dorny/test-reporter#229

It seems to have been the case for a while, but for some reason missing permissions are now breaking the build. It might be related to the other weird issues we've seen with Dependabot commits recently.

PR added the permission to mangrove.js, this commit adds it the rest of the jobs.