feat(addresses): Use ERC20 instance IDs
Instead of the ambiguous token symbols use the instance ID's assigned by context-addresses.
This makes all token instances available and enforces clarity on which ERC20 instance one is interacting with.

To ease the transition, the default instances are still available using just the token symbol as identifier.
However, use of this is deprecated.