ci: Do not run the `check` job if workflow is cancelled
Experiment using `!cancelled()` instead of `always()` as the condition for the `check` job in an attempt to avoid cancelled workflow runs blocking dependabot PR's from merging.

The problem appears to be that the `check` job runs and fails if the workflow is cancelled.
Workflow runs are often cancelled automatically when a newer run is triggered. In this case, the branch protection and dependabot should only consider the latest workflow run, not cancelled ones.