feat: Use network names from context-addresses
These network names coincide with the network names from the previous code, but adds more + constitutes a single source of truth for Mangrove.

At some point, mangrove.js (and the rest of Mangrove's code) should move to only using chain IDs as network names are unstable and non-canonical in the ecosystem.